pyproject.toml
[project]
name = "ai-agent-demo-02"
version = "0.1.0"
description = "Agentic AI: Python coding assistant"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
